feat: Subcommand to trim comparison files (#93)
Same values are considered extraneous if they cannot be reached by
moving up from a Different value in a number of steps defined in the
passed --level. Siblings of non-extraneous values are also kept.

For example, if passed --level 2, then all Same values that are not
siblings of or values than can be reached by moving up 2 or less steps
from a Different value would be omitted.